Allow Python to build and run Erlang scripts
Sends the necessary scipts to Erlang to be compiled and run. The compile
option returns an identifier for the code that can be later passed to
the run function. Alternatively, the script itself can be passed
directly to the run function.

Scripts to be run should mimic a escript with no header. Ie, they must
define a main/1 function that will be called with the provided argument.
The result of the main/1 function is returned to Python. An example
script might look something like:

    main(_) ->
        {ok, erlang:now()}.

Which would return the current timestamp from Erlang.